requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLO
ultralytics>=8.0.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# OCR
easyocr>=1.7.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress & Logging
tqdm>=4.65.0
loguru>=0.7.0

# Data Augmentation
albumentations>=1.3.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
pathlib>=1.0.1

# Video Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Metrics
scikit-image>=0.21.0

# For DINO
timm>=0.9.0

# Jupyter (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0
